+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Updated dependencies and configured new rules
+- Changed the array-element-newline options to allow multiline
 
 ## [1.0.0-alpha.1] - 2021-01-27